requirements.txt
caer
canaro
numpy
opencv-python
matplotlib
tensorflow
scikit-learn
keras
requests-html
beautifulsoup4